requirements/base.txt
appnope==0.1.0
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.1.2
Django==1.11
django-appconf==1.0.2
django-environ==0.4.4
django-rest-swagger==2.1.2
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
idna==2.6
ipdb==0.10.3
ipython==6.2.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.11.0
Jinja2==2.10
Markdown==2.6.9
MarkupSafe==1.0
mysqlclient==1.3.12
openapi-codec==1.3.2
parso==0.1.0
pexpect==4.3.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
psycopg2==2.7.1
ptyprocess==0.5.2
Pygments==2.2.0
PyJWT==1.5.3
pytz==2017.3
requests==2.18.4
simplegeneric==0.8.1
simplejson==3.13.2
six==1.11.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
wcwidth==0.1.7
